Changed class name
